+++ src/main/java/github/anandb/netbeans/ui/UIUtils.java
@@ -96,7 +96,7 @@ public class UIUtils {
     private static final Pattern CELL_SPLIT = Pattern.compile("(?<!\\\\)\\|");
 
     public static JButton createToolbarButton(String iconName, String toolTip, ActionListener l) {
-        return createToolbarButton(iconName, 28, toolTip, l);
+        return createToolbarButton(iconName, PluginSettings.getToolbarIconSize(), toolTip, l);
     }
 
     public static JButton createToolbarButton(String iconName, int iconSize, String toolTip, ActionListener l) {